requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
scikit-learn==1.3.2
scikit-image==0.22.0
joblib==1.3.2
Pillow==10.0.1
tqdm==4.66.1

# PyTorch - Cần cho CNN
# Cài đặt theo hướng dẫn tại: https://pytorch.org/get-started/locally/
# CPU only:
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
# GPU (CUDA 12.4):
#   pip install torch --index-url https://download.pytorch.org/whl/cu124
torch>=2.0.0
